chipsec: Bump to v1.3.6rc1
It has been a while since the last official release of CHIPSEC. For now,
bump to v1.3.6rc1. This will bring miscellaneous fixes and new features
such as the new module with checks for the mitigations for Spectre
variant 2.

This change does no require updates in any of the patches we carry as
no patch fuzz warnings are observed.

Signed-off-by: Ricardo Neri <ricardo.neri-calderon at linux.intel.com>
Signed-off-by: Megha Dey <megha.dey@linux.intel.com>