sbsigntool: use mainline git repo
Until now, we were using a fork of sbsigntool git repo. This was based
on a older version of sbsigntool mainline repo. It is better for
us to switch to the mainline repo so that it uses updated dependencies.
For example, the older sbsigntool was compatible with openssl10 and not
the newer openssh1.1.

We have to accordingly rebase the patch files associated with this
recipe.

Changes V1->V2:
Update the sbsigntool repo instead of updating dependencies of the older
sbsigntool repo.

Signed-off-by: Ricardo Neri <ricardo.neri-calderon at linux.intel.com>
Signed-off-by: Megha Dey <megha.dey@linux.intel.com>